start to make sh code reusable.
hope the configure doesn't blow up, locally tested.

this now requires readlink(1) and so far only serves
the purpose to get to call python indepent of setting
a python version in a Makefile for which with the current
layout we'd have to run configure which would require
python. Or we just record the location or name of the
python executable in config.mk...